@@ -1,6 +1,1 @@
-module.exports = {
-  plugins: {
-    tailwindcss: {},
-    autoprefixer: {},
-  },
-};
+export default { plugins: { tailwindcss: {}, autoprefixer: {} } };
